Update Snowman Fairmath language
Update the example of Fairmath in Snowman to use the word "increase" and "decrease" to avoid confusion over their operations as suggested by Greyelf (https://github.com/iftechfoundation/twine-cookbook/issues/8#issuecomment-354496585).

Adjusted wording within example to highlight the percentage change and not simply use of addition or subtraction to change the numbers.