Version 3
Editing pass for Version 3. Changes include technical editing pass changes to wording and punctuation on example pages; the removal of the Organization page (added to the GitHib wiki); a re-write of the README file; addition of the JavaScript page under Common Terms; change to ```window.setup``` for all examples except for those using shims (Geolocation and Story Saving in Snowman).